Import of a few assets, new folder in "Downloaded Assets"
I uploaded a few assets that i thought we could use for our project. so far almost only assets for "makeplayerfaster" powerup. Also created a folder by the same name under the downloaded asset folder inorder to increase clarity. This push should not break/manipulate anything in respect to the functionality of the game.