Got Speedup powerup working (with issue)
Issue:
increase of speed is adjustable on the game object but return to previous speed is not relative but fixed. So as of now the speed increase is + 5 (adjustable) and after _poweuptimeout it is -5 (not adjustable).
because the relevant variable is not avaiblabe in the relevant function. Not a major issue imo opinion because we can just make it a fixed speedup of +5  .