Edited Speedup - issue remains
I tried making the increased spead variable public in order to have access to it from the player script but it didnt work it takes the hard coded int value from the script and not the adjusted one from the unity editor.  its a working progress but main functionality works.