shield powerup
idea was to let an object spawn ontop of player that is semi transparent this object has no collider and nothing. upon powerup pickup we increse life by one and check in the damage function if Shield is on and if it is we destroy the shield and set the shield bool to false again.
basically addlife with a fancy bubble.